code cleanup (irrelevant method hop removed)
